Reworked... Everything
Didn't do a good job of doing commits on this one because it was basically
completely busted. Ran into issues with getting proper events and
timings, but this now fully works. It pulls a complete pair of build
orders from a replay, saves it as a json, and makes a folder
if you don't have one for your build orders.

Future changes will include supply counts, hopefully, and probably
save the seperate builds for player 1 and 2 into different files.