Adding Supply!!!
Had to rework several things, but added the ability to track Supply
counts for better build data. Needs to be tested a bit, but it seems
functional. Have to take the build data then sort it at the end in order
to have the correct order to place in supply