Adding Zerg Build Extractor
Running into several issues, but getting to a stopping place.
Have it mostly working now, need to be abel to figure out timings of
upgrades and then need to add buildings in, but otherwise getting close.

I realize that I need to refactor much of these scrips to include more
functions that will check the events because the if statements are getting
hard to follow.